Extract class_eval code to included module
